modified:   src/v2i-hub/MapPlugin/manifest.json 	modified:   src/v2i-hub/MapPlugin/src/MapPlugin.cpp 	new file:   src/v2i-hub/MapPlugin/src/MapPlugin.h
